credit-report-processor-gcf/requirements.txt
# Google Cloud Functions automatically provides Flask/Werkzeug for HTTP triggers
google-cloud-vision==3.4.5
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
supabase==2.3.4
psycopg2-binary==2.9.9
requests==2.31.0
functions-framework==3.5.0
